python-testing/pyproject.toml
[project]
name = "teamup-testing"
version = "0.1.0"
description = "Simple Python test suite for TeamUP PostgreSQL database"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0.27",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "faker>=23.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["teamup"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=teamup",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.coverage.run]
source = ["teamup"]
omit = ["*/tests/*", "*/__pycache__/*", "*/venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
